refactor(open-mpm): split tools/ batch 2 — ast_tools/tm_tools/git_tools
Split three oversized tool files into focused sub-module directories,
each under the 500-line cap, with thin mod.rs re-exporting the public
surface (factory fns + tool structs). No public API or behavior change.

- ast_tools.rs (801) -> ast_tools/{patch_store,get_symbol,edit_insert,
  import_validate,apply_patch,mod}.rs
- tm_tools.rs (731)  -> tm_tools/{helpers,query,lifecycle,control,mod}.rs
- git_tools.rs (698) -> git_tools/{helpers,inspect,branch,write,remote,mod}.rs

Part of #361.
